Enum hermit_abi::io::ErrorKind

#[non_exhaustive]
pub enum ErrorKind {

    AlreadyExists,
    NotSocket,
    NotFound,
    NotListening,
    InUse,
    ConnectionRefused,
    ConnectionReset,
    NotConnected,
    AddrInUse,
    AddrNotAvailable,
    WouldBlock,
    InvalidInput,
    InvalidData,
    TimedOut,
    WriteZero,
    Other,
    Unsupported,

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AlreadyExists

creating / adding failed

NotSocket

called function on invalid socket

NotFound

target address not found

NotListening

called accept on non-listening socket

InUse

socket is already used for another purpose

ConnectionRefused

connection refused by peer

ConnectionReset

connection reset by peer

NotConnected

tried to read from or write to unconnected socket

AddrInUse

address/port requested already in use

AddrNotAvailable

address/port requested not available

WouldBlock

action would block

InvalidInput

invalid input argument configuration (e.g. register event on wrong socket type)

InvalidData

data in input is invalid (e.g. not events specified on sys_register_event)

TimedOut

action timed out

WriteZero

tried to write 0 bytes

since a return value of zero from a write indicates an unwritable socket a zero write, for which a successful completion would return 0, is disallowed

Other

internal errors not directly addressable in user space

Unsupported

currently unsupported call